Implemented delete process and refactored code base
Refactoring code using locallib.php, adding delete process from external db data and as fallback from moodle db. Refactoring process for suspending users into normal user data update process
