+++ db/tasks.php
@@ -39,13 +39,6 @@ $tasks = [
         'disabled'=> 1, 
     ],
     [
-        'classname' => 'local_importpossehluser\task\disable_possehluser_cron',
-        'blocking' => 0,
-        'hour' => '1',
-        'day' => '*',
-        'disabled'=> 1, 
-    ],
-    [
         'classname' => 'local_importpossehluser\task\delete_possehluser_cron',
         'blocking' => 0,
         'hour' => '1',
