+++ version.php
@@ -25,7 +25,7 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->version   = 20220729472; // The current plugin version (Date: YYYYMMDDXX).
+$plugin->version   = 20220729473; // The current plugin version (Date: YYYYMMDDXX).
 $plugin->requires  = 2020041200; // Requires this Moodle version.
 $plugin->component = 'local_importpossehluser'; // Full name of the plugin (used for diagnostics)
